Add unit tests for OrderStockConversionService
- Implement tests for converting order items to stock items, including handling consumable, raw material, and manufactured products.
- Add tests for logging warnings for raw materials and breaking down manufactured products into their components.
- Include tests for stock removal and addition for completed and cancelled orders, respectively.
- Validate stock availability and summarize stock requirements for orders.
- Ensure proper handling of nested manufactured products and optimization of stock items by summing common components.

+++ larament/database/migrations/2025_07_07_000016_create_product_components_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
             $table->foreignId('component_id')->constrained('products')->onDelete('cascade');
-            $table->integer('quantity');
+            $table->double('quantity');
             $table->timestamps();
             $table->unique(['product_id', 'component_id']);
         });
