+++ larament/database/migrations/2025_07_07_000024_create_purchase_invoice_items_table.php
@@ -4,15 +4,14 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up(): void
     {
         Schema::create('purchase_invoice_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('purchase_invoice_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('price', 10, 2);
             $table->decimal('total', 12, 2);
             $table->timestamps();
