+++ larament/database/migrations/2025_07_07_000025_create_return_purchase_invoice_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('return_purchase_invoice_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity');
+            $table->decimal('quantity', 10, 2);
             $table->decimal('price', 10, 2);
             $table->decimal('total', 12, 2);
             $table->timestamps();
